+++ composer.json
@@ -36,7 +36,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Spatie\\TranslationLoader\\Test\\": "tests"
+            "Tests\\": "tests"
         }
     },
     "scripts": {
