+++ tests/Feature/DummyManagerTest.php
@@ -2,24 +2,11 @@
 
 declare(strict_types=1);
 
-use Spatie\TranslationLoader\Test\TestCase;
-use Spatie\TranslationLoader\Test\TranslationManagers\DummyManager;
-
-uses(TestCase::class);
+use Tests\Feature\TranslationManagers\DummyManager;
 
 beforeEach(function () {
     $this->app['config']->set('translation-loader.translation_manager', DummyManager::class);
-    //config()->set('translation-loader.translation_manager', DummyManager::class);
 });
-
-///**
-// * @param  Application  $app
-// */
-//function getEnvironmentSetUp($app): void
-//{
-//    getEnvironmentSetUp($app);
-//    $app['config']->set('translation-loader.translation_manager', DummyManager::class);
-//}
 
 it('allow to change translation manager', function () {
     expect($this->app['translation.loader'])->toBeInstanceOf(DummyManager::class);
