+++ tests/Feature/TranslationManagers/DummyManager.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Feature\TranslationManagers;
+
+use Spatie\TranslationLoader\TranslationLoaderManager;
+
+class DummyManager extends TranslationLoaderManager {}
