+++ tests/Fixtures/lang/en/file.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 return [
     'key' => 'en value',
     '404' => [
