+++ tests/Fixtures/lang/nl/file.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 return [
     'key' => 'nl value',
     '404' => [
