--- tests/TranslationManagers/DummyManager.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Spatie\TranslationLoader\Test\TranslationManagers;
-
-use Spatie\TranslationLoader\TranslationLoaderManager;
-
-class DummyManager extends TranslationLoaderManager
-{
-}
